Update for use with EMR-4.1.0 (Spark-1.5)
Pyspark 1.5.1 includes an important fix for serialization of named
tuples, so that gets pulled as part of the bootstrap process (though it
needs to replace a ZIP that's installed as part of spark-core (yum)
after the bootstrap actions complete).

EMR-4.1.0 supports variably-sized clusters and will begin processing as
soon as part of the cluster is available, allowing us to make better use
of reserved + spot instances.